+++ backend/src/main/java/com/example/airbnb/config/SwaggerConfig.java
@@ -1,0 +1,43 @@
+package com.example.airbnb.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+import java.util.HashSet;
+import java.util.Set;
+
+@EnableSwagger2
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public Docket api() {
+        ApiInfo apiInfo = new ApiInfoBuilder()
+                .title("Airbnb API")
+                .description("Spring API made by Bat")
+                .license("Apache License Version 2.0")
+                .version("0.1")
+                .build();
+
+        Set<String> responseContentType = new HashSet<>();
+        responseContentType.add("application/json;charset=UTF-8");
+
+        // http://localhost:8080/swagger-ui.html
+        return new Docket(DocumentationType.SWAGGER_2)
+                .groupName("airbnb")
+                .produces(responseContentType)
+                .apiInfo(apiInfo)
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("com.example.airbnb")) // NOTE: basic-error-controller 는 문서화 하지 않음
+                .paths(PathSelectors.ant("/**"))
+                .build();
+    }
+
+}
